refactor: move agent type to specified service type
