+++ src/core/agent/agent.types.ts
@@ -7,15 +7,12 @@ import { EventService } from "./services/eventService";
 import { IdentifierStorage } from "./records/identifierStorage";
 import { CredentialStorage } from "./records/credentialStorage";
 import { BasicStorage } from "./records/basicStorage";
+import { ConnectionHistoryType } from "./services/connection.types";
 
 enum ConnectionStatus {
   CONFIRMED = "confirmed",
   PENDING = "pending",
   ACCEPTED = "accepted",
-}
-
-enum ConnectionHistoryType {
-  CREDENTIAL_ACCEPTED,
 }
 
 interface ConnectionHistoryItem {
@@ -106,114 +103,15 @@ interface AgentServicesProps {
   credentialStorage: CredentialStorage;
 }
 
-enum MultiSigRoute {
-  ROT = "/multisig/rot",
-  ICP = "/multisig/icp",
-  IXN = "/multisig/ixn",
-}
-
 interface CreateIdentifierResult {
   signifyName: string;
   identifier: string;
-}
-
-interface KeriaContact {
-  alias: string;
-  id: string;
-  oobi: string;
-  challenges: string[];
-  wellKnowns: string[];
 }
 
 interface IdentifierResult {
   name: string;
   prefix: string;
   salty: any;
-}
-
-interface IdentifiersListResult {
-  aids: IdentifierResult[];
-  start: 0;
-  end: 0;
-  total: 0;
-}
-
-interface CreateMultisigExnPayload {
-  gid: string;
-  smids: any[];
-  rmids: any;
-  rstates: any;
-  name: string;
-}
-
-interface Aid {
-  name: string;
-  prefix: string;
-  salty: any;
-  transferable: boolean;
-  state: {
-    vn: number[];
-    i: string;
-    s: string;
-    p: string;
-    d: string;
-    f: string;
-    dt: string;
-    et: string;
-    kt: string;
-    k: string[];
-    nt: string;
-    n: string[];
-    bt: string;
-    b: string[];
-    c: string[];
-    ee: {
-      s: string;
-      d: string;
-      br: any[];
-      ba: any[];
-    };
-    di: string;
-  };
-  windexes: number[];
-}
-
-interface MultiSigExnMessage {
-  exn: {
-    v: string;
-    t: string;
-    d: string;
-    i: string;
-    p: string;
-    dt: string;
-    r: string;
-    q: any;
-    a: {
-      gid: string;
-      smids: string[];
-      rmids: string[];
-      rstates: Aid["state"][];
-      name: string;
-    };
-    e: {
-      icp: {
-        v: string;
-        t: string;
-        d: string;
-        i: string;
-        s: string;
-        kt: string;
-        k: string[];
-        nt: string;
-        n: string[];
-        bt: string;
-        b: string[];
-        c: any[];
-        a: any[];
-      };
-      d: string;
-    };
-  };
 }
 
 enum NotificationRoute {
@@ -224,12 +122,10 @@ enum NotificationRoute {
 
 export {
   ConnectionStatus,
-  ConnectionHistoryType,
   MiscRecordId,
   ConnectionEventTypes,
   AcdcEventTypes,
   NotificationRoute,
-  MultiSigRoute,
 };
 
 export type {
@@ -245,10 +141,5 @@ export type {
   KeriaNotificationMarker,
   AgentServicesProps,
   CreateIdentifierResult,
-  KeriaContact,
-  IdentifiersListResult,
-  CreateMultisigExnPayload,
-  MultiSigExnMessage,
-  Aid,
   IdentifierResult,
 };
