+++ src/core/agent/services/connection.types.ts
@@ -1,0 +1,15 @@
+interface KeriaContact {
+  alias: string;
+  id: string;
+  oobi: string;
+  challenges: string[];
+  wellKnowns: string[];
+}
+
+enum ConnectionHistoryType {
+  CREDENTIAL_ACCEPTED,
+}
+
+export { ConnectionHistoryType };
+
+export type { KeriaContact };
